# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/token-limit-configs": {
      "post": {
        "operationId": "createTokenLimitConfig",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenLimitConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a token-limit config for a subscription tier",
        "tags": [
          "Token Limit Config"
        ]
      },
      "get": {
        "operationId": "findAllTokenLimitConfigs",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all token-limit configs",
        "tags": [
          "Token Limit Config"
        ]
      }
    },
    "/token-limit-configs/subscription/{subscription}": {
      "get": {
        "operationId": "findTokenLimitConfigBySubscription",
        "parameters": [
          {
            "name": "subscription",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "free",
                "basic"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the token-limit config for a specific subscription tier",
        "tags": [
          "Token Limit Config"
        ]
      }
    },
    "/token-limit-configs/{id}": {
      "get": {
        "operationId": "findTokenLimitConfigById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a token-limit config by id",
        "tags": [
          "Token Limit Config"
        ]
      },
      "put": {
        "operationId": "updateTokenLimitConfig",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTokenLimitConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a token-limit config",
        "tags": [
          "Token Limit Config"
        ]
      },
      "delete": {
        "operationId": "deleteTokenLimitConfig",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a token-limit config",
        "tags": [
          "Token Limit Config"
        ]
      }
    },
...
    "/admin/users": {
      "get": {
        "operationId": "adminListUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all users",
        "tags": [
          "Admin"
        ]
      },
      "post": {
        "operationId": "adminCreateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "operationId": "adminGetUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a user by id",
        "tags": [
          "Admin"
        ]
      },
      "patch": {
        "operationId": "adminUpdateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "adminDeleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/users/{id}/reset-tokens": {
      "post": {
        "operationId": "adminResetUserTokens",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reset a user's token usage counter",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "LMStudio Wrapper API",
    "description": "REST API for LM Studio Wrapper",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTokenLimitConfigDto": {
        "type": "object",
        "properties": {
          "minutesTillReset": {
            "type": "number",
            "example": 1
          },
          "tokensPerInterval": {
            "type": "number",
            "example": 9000
          },
          "subscription": {
            "type": "string",
            "enum": [
              "free",
              "basic"
            ],
            "example": "basic"
          }
        },
        "required": [
          "minutesTillReset",
          "tokensPerInterval",
          "subscription"
        ]
      },
      "UpdateTokenLimitConfigDto": {
        "type": "object",
        "properties": {
          "minutesTillReset": {
            "type": "number",
            "example": 1
          },
          "tokensPerInterval": {
            "type": "number",
            "example": 9000
          },
          "subscription": {
            "type": "string",
            "enum": [
              "free",
              "basic"
            ],
            "example": "basic"
          }
        }
      },
...
      "CreateAdminUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "user"
            ],
            "default": "user"
          },
          "subscription": {
            "type": "string",
            "enum": [
              "free",
              "basic"
            ],
            "default": "free"
          },
          "isActivated": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "UpdateAdminUserDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "description": "New password — leave unset to keep the current one"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "user"
            ]
          },
          "subscription": {
            "type": "string",
            "enum": [
              "free",
              "basic"
            ]
          },
          "isActivated": {
            "type": "boolean"
          }
        }
      }
    }
  }
}